move masks to module; bun-types; tests
