+++ package.json
@@ -1,62 +1,62 @@
 {
-	"name": "chessalyzer.js",
-	"version": "3.0.6",
-	"description": "A node.js library for batch analyzing chess games.",
-	"author": "Yannik Schröder",
-	"type": "module",
-	"main": "lib/chessalyzer.js",
-	"types": "lib/chessalyzer.d.ts",
-	"exports": {
-		".": {
-			"types": "./lib/chessalyzer.d.ts",
-			"import": "./lib/chessalyzer.js"
-		}
-	},
-	"imports": {
-		"#bitboard": "./pkg/bitboard.js"
-	},
-	"files": [
-		"lib/chessalyzer.js",
-		"lib/chessalyzer.d.ts",
-		"lib/core",
-		"lib/tracker"
-	],
-	"engines": {
-		"node": ">=16.0.0"
-	},
-	"scripts": {
-		"build": "wasm-pack build --target nodejs && tsc && rollup --config",
-		"docs": "typedoc"
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/yschroe/chessalyzer.js"
-	},
-	"keywords": [
-		"chess",
-		"pgn",
-		"heatmap",
-		"batch",
-		"analysis"
-	],
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/yschroe/chessalyzer.js/issues"
-	},
-	"homepage": "https://github.com/yschroe/chessalyzer.js",
-	"devDependencies": {
-		"@types/bun": "^1.0.12",
-		"@types/node": "^20.11.28",
-		"@typescript-eslint/eslint-plugin": "^7.2.0",
-		"@typescript-eslint/parser": "^7.2.0",
-		"eslint": "^8.57.0",
-		"rollup": "^4.13.0",
-		"rollup-plugin-dts": "^6.1.0",
-		"typedoc": "^0.25.12",
-		"typedoc-plugin-merge-modules": "^5.1.0",
-		"typescript": "^5.4.2"
-	},
-	"dependencies": {
-		"chalk": "^5.3.0"
-	}
+  "name": "chessalyzer.js",
+  "version": "3.0.6",
+  "description": "A node.js library for batch analyzing chess games.",
+  "author": "Yannik Schröder",
+  "type": "module",
+  "main": "lib/chessalyzer.js",
+  "types": "lib/chessalyzer.d.ts",
+  "exports": {
+    ".": {
+      "types": "./lib/chessalyzer.d.ts",
+      "import": "./lib/chessalyzer.js"
+    }
+  },
+  "imports": {
+    "#bitboard": "./pkg/bitboard.js"
+  },
+  "files": [
+    "lib/chessalyzer.js",
+    "lib/chessalyzer.d.ts",
+    "lib/core",
+    "lib/tracker"
+  ],
+  "engines": {
+    "node": ">=16.0.0"
+  },
+  "scripts": {
+    "build": "wasm-pack build --target nodejs && tsc && rollup --config",
+    "docs": "typedoc"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/yschroe/chessalyzer.js"
+  },
+  "keywords": [
+    "chess",
+    "pgn",
+    "heatmap",
+    "batch",
+    "analysis"
+  ],
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/yschroe/chessalyzer.js/issues"
+  },
+  "homepage": "https://github.com/yschroe/chessalyzer.js",
+  "devDependencies": {
+    "@types/node": "^20.12.2",
+    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@typescript-eslint/parser": "^7.4.0",
+    "bun-types": "^1.0.36",
+    "eslint": "^8.57.0",
+    "rollup": "^4.13.2",
+    "rollup-plugin-dts": "^6.1.0",
+    "typedoc": "^0.25.12",
+    "typedoc-plugin-merge-modules": "^5.1.0",
+    "typescript": "^5.4.3"
+  },
+  "dependencies": {
+    "chalk": "^5.3.0"
+  }
 }
